requirements.txt
requirements.txtcertifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dictdiffer==0.8.1
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
viptela==0.2.4.dev2
